requirements.txt
plexapi
ytmusicapi
python-dotenv
questionary
colorama
tqdm